Try to prevent browsers from caching the log file.
This sends <meta> tags that are supposed to turn off caching in
browsers.  However, I'm not 100% convinced it works.  When I view the
source in a browser, the tags are missing.  I can't tell if it's
because it's really not working as expected, or it's working but the
browser is simply not displaying these tags in source view because
content is dynamically generated.